+++ src/pooldose/__init__.py
@@ -1,5 +1,5 @@
 """Async API client for SEKO Pooldose."""
 from .client import PooldoseClient
 
-__version__ = "0.7.0"
+__version__ = "0.7.5"
 __all__ = ["PooldoseClient"]